requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
email-validator
PyJWT[crypto]
httpx
pydantic
pydantic-settings
structlog
